migrate some method definition
